backend/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0

# Celery
celery==5.3.4
redis==5.0.1

# LangChain
langchain==0.1.0
langchain-openai==0.0.5
langchain-google-genai==1.0.0

# Database
supabase==2.0.0

# HTTP Client
httpx==0.25.0

# Crawler
beautifulsoup4==4.12.3

# Utilities
python-dotenv==1.0.0

# Logging (optional)
structlog==24.1.0
